Added Lap Time system + Terrain Sensing + Bobbing Animation
User can now race around the track, with the best Lap time being kept
in the top right of the window.
The game now sees what type of terrain the user is above and will
change linear drag and speed based on it to simulate rougher terrain.
Added a bobbing animation for future submarine sprites.